Refactor generate_bubbles.py into modular architecture
Split monolithic generate_bubbles.py into focused modules:

- boundaries.py: Boundary data operations, file downloads, output setup
- bubble_generation.py: Core bubble generation algorithm and geometry utilities
- analysis.py: Statistics calculation, coverage analysis, and visualization
- main.py: Application coordination, CLI interface, and processing pipeline

Updated bubble_experimentation.ipynb imports to work with new module structure.

This is a strict refactor preserving identical functionality while improving
code organization, maintainability, and testability. All function signatures,
return values, and behavior remain unchanged.
